Seperate picture compoent from home component
