+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/elle/turing/mod-3/project/purrfect-status/src/index.js":"1","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/App/index.js":"2","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Home/index.js":"3","/Users/elle/turing/mod-3/project/purrfect-status/src/apiCalls.js":"4","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Form/index.js":"5"},{"size":276,"mtime":1610682335093,"results":"6","hashOfConfig":"7"},{"size":166,"mtime":1610675676838,"results":"8","hashOfConfig":"7"},{"size":1708,"mtime":1610688166629,"results":"9","hashOfConfig":"7"},{"size":336,"mtime":1610667537830,"results":"10","hashOfConfig":"7"},{"size":1466,"mtime":1610683671626,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1swfd7p",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/elle/turing/mod-3/project/purrfect-status/src/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/App/index.js",[],["23","24"],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Home/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/apiCalls.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Form/index.js",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/elle/turing/mod-3/project/purrfect-status/src/index.js":"1","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/App/index.js":"2","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Home/index.js":"3","/Users/elle/turing/mod-3/project/purrfect-status/src/apiCalls.js":"4","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Form/index.js":"5","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Picture/index.js":"6"},{"size":276,"mtime":1610736339811,"results":"7","hashOfConfig":"8"},{"size":166,"mtime":1610736339806,"results":"9","hashOfConfig":"8"},{"size":669,"mtime":1610737390683,"results":"10","hashOfConfig":"8"},{"size":336,"mtime":1610736339811,"results":"11","hashOfConfig":"8"},{"size":1466,"mtime":1610736339808,"results":"12","hashOfConfig":"8"},{"size":1429,"mtime":1610737425755,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1swfd7p",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/elle/turing/mod-3/project/purrfect-status/src/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/App/index.js",[],["27","28"],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Home/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/apiCalls.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Form/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Picture/index.js",["29"],{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","severity":1,"message":"35","line":17,"column":47,"nodeType":"36","endLine":17,"endColumn":49,"suggestions":"37"},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'generateRandomCatPicture'. Either include it or remove the dependency array.","ArrayExpression",["40"],"no-global-assign","no-unsafe-negation",{"desc":"41","fix":"42"},"Update the dependencies array to be: [generateRandomCatPicture]",{"range":"43","text":"44"},[525,527],"[generateRandomCatPicture]"]
