+++ src/Components/Home/index.js
@@ -1,26 +1,14 @@
-import React, { useState, useEffect } from 'react';
-import { getCatPicture } from '../../apiCalls';
+import React, { useState } from 'react';
+import Picture from '../Picture';
 import Form from '../Form';
 import './Home.scss'
-import { IoPawSharp } from 'react-icons/io5';
 
 const Home = () => {
-  const [pictureUrl, setPictureUrl] = useState('');
   const [userInput, setUserInput] = useState(null);
-  const [isFavorite, setIsFavorite] = useState(false);
-
-  const generateRandomCatPicture = () => {
-    getCatPicture().then(data =>{
-      setPictureUrl(data[0].url);
-      setUserInput(null);
-    })
-  }
 
   const getUserInput = newInput => {
     setUserInput(newInput);
   }
-
-  useEffect(() => generateRandomCatPicture(), [])
   
   return (
     <section className="home-page">
@@ -29,31 +17,7 @@ const Home = () => {
         <button>Status Pawtrait</button>
       </nav>
       <section className="display-picture-form-section">
-        <section className="picture-section">
-          <div 
-            className="cat-picture-wrapper" 
-            style={{
-              backgroundImage: `url(${pictureUrl})`,
-              backgroundPosition: "center",
-              backgroundSize: "cover",
-              backgroundRepeat: "no-repeat"
-            }}>
-            <div className="icon-wrapper">
-              {userInput && <IoPawSharp 
-                className="icon" 
-                onClick={() => setIsFavorite(prevStatus => !prevStatus)} 
-                style={{color: isFavorite ? '#177fcc' : 'white' }}
-              />}
-            </div>
-            <figure className="dispay-inputs-wrapper">
-              <p>{userInput && userInput.statusCode}</p>
-              <p>{userInput && userInput.explaination}</p>
-            </figure>
-          </div>
-          <div>
-            <button onClick={generateRandomCatPicture}>Not Purrfect</button>
-          </div>
-        </section>
+        <Picture setUserInput={setUserInput} userInput={userInput}/>
         <Form getUserInput={getUserInput}/>
       </section>
     </section>
